+++ Sagit/Category/UI/STUIButton.h
@@ -20,7 +20,7 @@
 -(UIButton*)title:(NSString*)title forState:(UIControlState)state;
 -(UIButton*)titleColor:(id)colorOrHex;
 -(UIButton*)titleColor:(id)colorOrHex forState:(UIControlState)state;
--(UIButton*)titleFont:(NSInteger)px;
+-(UIButton*)titleFont:(CGFloat)px;
 -(UIButton*)adjustsImageWhenHighlighted:(BOOL)yesNo;
 //!当button在动态设置文字或图片之后，宽度自适应
 -(UIButton*)stWidthToFit;
